move Num and impls to int and float
